Crate jrsonnet_parser

Re-exports

pub use peg;

Structs

ArgsDesc

AssertStmt

CodeLocation

ExprLocation

file, begin offset, end offset

FieldMember

ForSpecData

IStr

Interned string

IfSpecData

IndexPart

LocExpr

Holds AST expression and its location in source file

ObjComp

Param

name, default value

ParamsDesc

Defined function parameters

ParserSettings

SliceDesc

Source

Either real file, or virtual Hash of FileName always have same value as raw Path, to make it possible to use with raw_entry_mut

SourceDirectory

Represents path to the directory on the disk

SourceFile

Represents path to the file on the disk Directories shouldn’t be put here, as resolution for files differs from resolution for directories:

SourcePath

Represents location of a file

SourceVirtual

Represents virtual file, whose are located in memory, and shouldn’t be cached

Enums

BinaryOpType

BindSpec

CompSpec

Destruct

DestructRest

Expr

Syntax base

FieldName

LiteralType

Member

ObjBody

UnaryOpType

Visibility

Traits

SourcePathT

Functions

parse

string_to_expr

Used for importstr values

Type Aliases

ParseError